Make integer keys actually serialize as integer keys.
Object.keys() always returns a array of USVStrings which makes
the serialization serialize integer keys as string keys. This has
interoperability problems with other implementations that actually
serialize integer keys as integer keys.

An unfortunate side effect of this is that string keys which are
parseable as numbers will be serialized as numeric keys. This has a
positive side effect of reducing the final serialized output, and
works fine for us, but may not work for everyone.